Bugfix: Fixed Example projects so they work+ added ui improvement to help users. (#22)
bugfix: Removed Client Secret checking validation as it was impossible to run the app whilst leaving the client secret blank.  It would just re-open the validation sheet.
feature: added placeholder text to tell the user that they should leave the client secret blank. I left the fields in, just in case they are going to be used for other projects/ in the future.
admin: Added comments to state that the client secret should be blank so its understood easily by developers trying to use this.

Hope these improvements help -
Tested ok on latest Xcode 7.3.1