Handle errors coming back in an array.
Some servers version of HTTP Streaming can come into play for a GET request
if it the result was streamable, so we need deal with arrays of errors.

It would appear the only way this can happen is for an array of one error,
so always extract the first item, but debug log if there was more than
one as a hint that this might need better/more complete support in the
future.

Let Xcode 8.2 update the project/schemes while at it.