Support CLANG_WARN_OBJC_IMPLICIT_RETAIN_SELF.
CocoaPods is moving to enable CLANG_WARN_OBJC_IMPLICIT_RETAIN_SELF by
default, so enabled it in the project and update the code accordingly.

Blocks within the library implementation direct touch ivars back properties
as part of updating state with doing the work the public access to the
property requires, so use `self->` notation to make it an explicit
capture of `self` instead of an implicit one.

Looks like some of the Deps can't current support enabling this, so
leave it off for now.