Remove the support for service:version pairs on the commandline.
Everything in directory uses an url on another host for the discovery document,
so it doesn't make sense to support just taking a service:version and fetching
the document from the discovery service.

This as a side effect of making the support for skipping services a little odd
so change that to use an explicit --skip SERVICE argument instead.

Update the generation script to use the new syntax for skipping things.

Add an explicit check for files existing so there is a more obvious message
rather then that NSURL error that gets reported otherwise.