Honor deprecation on Resources.
- Track if a resource has a deprecated parent.
- Track if a method belongs to a resource with a deprecated parent.
- Update all checks for deprecated methods to see if they have a deprecated
  parent resources.

The current discovery data only shows one resource marked as deprecated and all
of the method defined off it are also deprecated, so there are no generation
changes with this change.

But this updated logic change is correct and ensure if the methods were marked
deprecated the marking of the resource as such would do the correct things.